Full repository import #11689

+++ modules/lib/lib-export/src/test/java/com/enonic/xp/lib/export/FunctionBasedNodeImportListenerTest.java
@@ -1,6 +1,6 @@
 package com.enonic.xp.lib.export;
 
-import java.util.concurrent.atomic.AtomicLong;
+import java.util.concurrent.atomic.AtomicInteger;
 
 import org.junit.jupiter.api.Test;
 
@@ -12,9 +12,9 @@ class FunctionBasedNodeImportListenerTest
     @Test
     void call_functions()
     {
-        AtomicLong nodesResolved = new AtomicLong();
-        AtomicLong nodesImported = new AtomicLong();
-        AtomicLong nodesSkipped = new AtomicLong();
+        AtomicInteger nodesResolved = new AtomicInteger();
+        AtomicInteger nodesImported = new AtomicInteger();
+        AtomicInteger nodesSkipped = new AtomicInteger();
 
         final FunctionBasedNodeImportListener functionBasedNodeImportListener = new FunctionBasedNodeImportListener( i -> {
             nodesImported.addAndGet( i );
